multiplication

+++ 0x04-more_functions_nested_loops/2-mul.c
@@ -9,5 +9,5 @@
 
 int mul(int a, int b)
 {
-	return ((long long) a * b);
+	return (a * b);
 }
